A better way to inject the hyperion views into the layout hierarchy.
Many issues we've run into with accessibility, focus, etc. are due to
hyperion injecting views into the layout hierarchy even when it's not
shown. To fix this, no views are added until hyperion is shown. And when
it is, the additional views are added to the window's content view
keeping the app content in place.

Behavior changes:
1) A plugin's view is not always guaranteed to be attached to the
layout, as it's attached when the menu is shown and detached when the
menu is hidden. This can happen multiple times between when a plugin is
created and destroyed.
2) `MeasurementHelper#getScreenLocation` coordinates may not line up
with with the content view when used in an overlay. This is because the
overlay is now placed at the same level as the content view so it might
be moved down a bit to make room for the status bar etc. You can use
MeasurementHelper#getContentRootLocation` instead which will get you
coordinates relative to the content root (and the overlay).

Known issues:
There a some graphical glitches in the sample app that needs to be
investigated.